my_team/pyproject.toml
[tool.poetry]
name = "my-team"
version = "0.1.0"
description = ""
authors = ["Mihail Pugachev <xd393727@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
unicorn = {extras = ["standart"], version = "^2.0.1.post1"}
alembic = "^1.13.1"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
aiosqlite = "^0.19.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
